# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khata Payment Ledger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LedgerProfile {
  phone?: string | null;
  upiId?: string | null;
  bankAccount?: string | null;
  ifsc?: string | null;
  notes?: string | null;
}

export interface Ledger {
  id: string;
  name: string;
  fifoEnabled: boolean;
  profile?: LedgerProfile;
  createdAt: string;
  updatedAt: string;
}

export type LedgerWithBalance = Ledger & {
  balance: number;
  totalCredit: number;
  totalDebit: number;
  entryCount: number;
};

export interface CreateLedgerInput {
  name: string;
  fifoEnabled?: boolean;
  profile?: LedgerProfile;
}

export interface UpdateLedgerInput {
  name?: string;
  fifoEnabled?: boolean;
  profile?: LedgerProfile;
}

export type EntryType = (typeof EntryType)[keyof typeof EntryType];

export const EntryType = {
  credit: "credit",
  debit: "debit",
  transfer: "transfer",
} as const;

export interface Entry {
  id: string;
  ledgerId: string;
  date: string;
  remark: string;
  amount: number;
  type: EntryType;
  transferToLedgerId?: string | null;
  transferToLedgerName?: string | null;
  linkedTransferId?: string | null;
  fifoRemaining?: number | null;
  balance: number;
  createdAt: string;
}

export type CreateEntryInputType =
  (typeof CreateEntryInputType)[keyof typeof CreateEntryInputType];

export const CreateEntryInputType = {
  credit: "credit",
  debit: "debit",
  transfer: "transfer",
} as const;

export interface CreateEntryInput {
  date: string;
  remark: string;
  amount: number;
  type: CreateEntryInputType;
  transferToLedgerId?: string | null;
}

export interface UpdateEntryInput {
  date?: string;
  remark?: string;
  amount?: number;
}

export interface LedgerSummary {
  totalCredit: number;
  totalDebit: number;
  balance: number;
  entryCount: number;
  openingBalance: number;
}

export interface FifoBill {
  entryId: string;
  date: string;
  remark: string;
  originalAmount: number;
  paidAmount: number;
  remainingAmount: number;
  isFullyPaid: boolean;
}

export interface FifoStatus {
  enabled: boolean;
  oldestUnpaidBill?: FifoBill | null;
  bills: FifoBill[];
  totalOutstanding: number;
}

export type OverviewRecentActivityItem = {
  ledgerId: string;
  ledgerName: string;
  entryId: string;
  date: string;
  remark: string;
  amount: number;
  type: string;
};

export interface Overview {
  totalLedgers: number;
  totalOutstanding: number;
  ledgersOwingMe: number;
  ledgersIOweThem: number;
  recentActivity: OverviewRecentActivityItem[];
}

export type GetEntriesParams = {
  startDate?: string;
  endDate?: string;
};

export type GetLedgerSummaryParams = {
  startDate?: string;
  endDate?: string;
};