# lib/api-zod/src/generated/types/createEntryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khata Payment Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateEntryInputType } from "./createEntryInputType";

export interface CreateEntryInput {
  date: Date;
  remark: string;
  amount: number;
  type: CreateEntryInputType;
  transferToLedgerId?: string | null;
}