# lib/api-zod/src/generated/types/entry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khata Payment Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { EntryType } from "./entryType";

export interface Entry {
  id: string;
  ledgerId: string;
  date: Date;
  remark: string;
  amount: number;
  type: EntryType;
  transferToLedgerId?: string | null;
  transferToLedgerName?: string | null;
  linkedTransferId?: string | null;
  fifoRemaining?: number | null;
  balance: number;
  createdAt: Date;
}