# lib/api-zod/src/generated/types/fifoBill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khata Payment Ledger API
 * OpenAPI spec version: 0.1.0
 */

export interface FifoBill {
  entryId: string;
  date: Date;
  remark: string;
  originalAmount: number;
  paidAmount: number;
  remainingAmount: number;
  isFullyPaid: boolean;
}