# lib/api-zod/src/generated/types/fifoStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khata Payment Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { FifoBill } from "./fifoBill";

export interface FifoStatus {
  enabled: boolean;
  oldestUnpaidBill?: FifoBill | null;
  bills: FifoBill[];
  totalOutstanding: number;
}