# lib/api-zod/src/generated/types/ledgerProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khata Payment Ledger API
 * OpenAPI spec version: 0.1.0
 */

export interface LedgerProfile {
  phone?: string | null;
  upiId?: string | null;
  bankAccount?: string | null;
  ifsc?: string | null;
  notes?: string | null;
}