# lib/api-zod/src/generated/types/ledgerSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khata Payment Ledger API
 * OpenAPI spec version: 0.1.0
 */

export interface LedgerSummary {
  totalCredit: number;
  totalDebit: number;
  balance: number;
  entryCount: number;
  openingBalance: number;
}