# lib/api-zod/src/generated/types/ledgerWithBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khata Payment Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { Ledger } from "./ledger";

export type LedgerWithBalance = Ledger & {
  balance: number;
  totalCredit: number;
  totalDebit: number;
  entryCount: number;
};