# lib/api-zod/src/generated/types/overview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khata Payment Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { OverviewRecentActivityItem } from "./overviewRecentActivityItem";

export interface Overview {
  totalLedgers: number;
  totalOutstanding: number;
  ledgersOwingMe: number;
  ledgersIOweThem: number;
  recentActivity: OverviewRecentActivityItem[];
}