# lib/api-zod/src/generated/types/updateEntryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khata Payment Ledger API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateEntryInput {
  date?: Date;
  remark?: string;
  amount?: number;
}