# lib/api-zod/src/generated/types/updateLedgerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khata Payment Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { LedgerProfile } from "./ledgerProfile";

export interface UpdateLedgerInput {
  name?: string;
  fifoEnabled?: boolean;
  profile?: LedgerProfile;
}